Refactored the ICMPEchoPacket classes some so that they can be reused with the PingReply class.

+++ core/jicmpv6/src/main/java/org/opennms/jicmp/ip/ICMPEchoPacket.java
@@ -61,7 +61,7 @@ public class ICMPEchoPacket extends ICMPPacket {
         return m_packetData.limit();
     }
 
-    public long getIdentifier() {
+    public int getIdentifier() {
         return getUnsignedShort(4);
     }
