+++ core/jicmpv6/src/main/java/org/opennms/jicmp/ipv6/ICMPv6EchoPacket.java
@@ -57,7 +57,7 @@ public class ICMPv6EchoPacket extends ICMPv6Packet {
         return getContentBuffer().array();
     }
     
-    public long getIdentifier() {
+    public int getIdentifier() {
         return getUnsignedShort(4);
     }
